requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
datasets==4.8.4
dill==0.4.1
distro==1.9.0
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.13
jiter==0.14.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
numpy==2.4.4
openai==2.32.0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
propcache==0.4.1
pyarrow==24.0.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
